requirements.txt
Django>=4.0
psycopg2-binary
gunicorn
Pillow           # 👈 جدید: برای مدیریت فایل‌ها و تصاویر
djangorestframework # 👈 جدید: برای ایجاد REST API
Pillow
python-dotenv

